+++ test/dev/se_test.js
@@ -1680,6 +1680,7 @@ const options1 = {
 			$.plugins.slashCommand.open(block);
 		},
 	},
+	textDirection: 'rtl',
 	slashCommand: {
 		// triggerChar: '/',
 		limitSize: 17,